feat: convert input type blocks to one block with dropdown
