+++ examples/developer-tools/src/blocks/index.ts
@@ -6,7 +6,7 @@
 
 import {factoryBase} from './factory_base';
 import * as Blockly from 'blockly';
-import {inputDummy, inputStatement, inputValue, inputEndRow} from './inputs';
+import {input} from './input';
 
 // import for side effects for now
 import '../output-generators/factory_base';
@@ -14,10 +14,7 @@ import '../output-generators/factory_base';
 export const registerAllBlocks = function() {
   Blockly.common.defineBlocks({
     'factory_base': factoryBase,
-    'input_value': inputValue,
-    'input_statement': inputStatement,
-    'input_dummy': inputDummy,
-    'input_end_row': inputEndRow,
+    'input': input,
   });
 };
